# api/handlers_git.go
package api

import (
	"fmt"
	"net/http"
	"strconv"

	"github.com/caiatech/govc/pkg/object"
	"github.com/gin-gonic/gin"
)

// addFile adds a file to the repository
// @Summary Add a file to repository
// @Description Writes a file to the working directory and stages it for commit
// @Tags Git Operations
// @Accept json
// @Produce json
// @Param repo_id path string true "Repository ID"
// @Param request body AddFileRequest true "File addition request"
// @Success 200 {object} SuccessResponse "File added successfully"
// @Failure 400 {object} ErrorResponse "Invalid request data"
// @Failure 404 {object} ErrorResponse "Repository not found"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /repositories/{repo_id}/files [post]
func (s *Server) addFile(c *gin.Context) {
	repoID := c.Param("repo_id")
	repo, err := s.getRepository(repoID)
	if err != nil {
		c.JSON(http.StatusNotFound, ErrorResponse{
			Error: err.Error(),
			Code:  "REPO_NOT_FOUND",
		})
		return
	}

	var req AddFileRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, ErrorResponse{
			Error: err.Error(),
			Code:  "INVALID_REQUEST",
		})
		return
	}

	// First write the content to the worktree
	if err := repo.WriteFile(req.Path, []byte(req.Content)); err != nil {
		c.JSON(http.StatusInternalServerError, ErrorResponse{
			Error: fmt.Sprintf("failed to write file: %v", err),
			Code:  "WRITE_FAILED",
		})
		return
	}

	// Then add it to staging
	if err := repo.Add(req.Path); err != nil {
		c.JSON(http.StatusInternalServerError, ErrorResponse{
			Error: fmt.Sprintf("failed to add file to staging: %v", err),
			Code:  "ADD_FAILED",
		})
		return
	}

	c.JSON(http.StatusOK, SuccessResponse{
		Status:  "added",
		Message: fmt.Sprintf("file %s added to staging", req.Path),
	})
}

// commit creates a new commit
// @Summary Create a new commit
// @Description Creates a new commit with staged changes
// @Tags Git Operations
// @Accept json
// @Produce json
// @Param repo_id path string true "Repository ID"
// @Param request body CommitRequest true "Commit request"
// @Success 201 {object} CommitResponse "Commit created successfully"
// @Failure 400 {object} ErrorResponse "Invalid request data"
// @Failure 404 {object} ErrorResponse "Repository not found"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /repositories/{repo_id}/commits [post]
func (s *Server) commit(c *gin.Context) {
	repoID := c.Param("repo_id")
	repo, err := s.getRepository(repoID)
	if err != nil {
		c.JSON(http.StatusNotFound, ErrorResponse{
			Error: err.Error(),
			Code:  "REPO_NOT_FOUND",
		})
		return
	}

	var req CommitRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, ErrorResponse{
			Error: err.Error(),
			Code:  "INVALID_REQUEST",
		})
		return
	}

	// Set author if provided
	if req.Author != "" {
		repo.SetConfig("user.name", req.Author)
	}
	if req.Email != "" {
		repo.SetConfig("user.email", req.Email)
	}

	// Create commit
	commit, err := repo.Commit(req.Message)
	if err != nil {
		c.JSON(http.StatusInternalServerError, ErrorResponse{
			Error: fmt.Sprintf("failed to create commit: %v", err),
			Code:  "COMMIT_FAILED",
		})
		return
	}

	c.JSON(http.StatusCreated, CommitResponse{
		Hash:      commit.Hash(),
		Message:   commit.Message,
		Author:    commit.Author.Name,
		Email:     commit.Author.Email,
		Timestamp: commit.Author.Time,
		Parent:    commit.ParentHash,
	})
}

// getLog returns commit history
// @Summary Get commit history
// @Description Retrieves the commit history for a repository
// @Tags Git Operations
// @Produce json
// @Param repo_id path string true "Repository ID"
// @Param limit query int false "Maximum number of commits to return (default: 50, max: 10000)"
// @Success 200 {object} object{commits=[]CommitResponse,count=int} "Commit history"
// @Failure 400 {object} ErrorResponse "Invalid request parameters"
// @Failure 404 {object} ErrorResponse "Repository not found"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /repositories/{repo_id}/commits [get]
func (s *Server) getLog(c *gin.Context) {
	repoID := c.Param("repo_id")
	repo, err := s.getRepository(repoID)
	if err != nil {
		c.JSON(http.StatusNotFound, ErrorResponse{
			Error: err.Error(),
			Code:  "REPO_NOT_FOUND",
		})
		return
	}

	// Parse query parameters
	limit := 50
	if l := c.Query("limit"); l != "" {
		parsed, err := strconv.Atoi(l)
		if err != nil {
			c.JSON(http.StatusBadRequest, ErrorResponse{
				Error: fmt.Sprintf("invalid limit parameter: %s", l),
				Code:  "INVALID_LIMIT",
			})
			return
		}
		if parsed <= 0 {
			c.JSON(http.StatusBadRequest, ErrorResponse{
				Error: fmt.Sprintf("limit must be positive: %d", parsed),
				Code:  "INVALID_LIMIT",
			})
			return
		}
		// Cap the limit to prevent excessive memory usage
		if parsed > 10000 {
			parsed = 10000
		}
		limit = parsed
	}

	// Get commits
	commits, err := repo.Log(limit)
	if err != nil {
		// Check if it's just an empty repository
		if err.Error() == "object not found: " || err.Error() == "reference not found" {
			// Empty repository, return empty list
			c.JSON(http.StatusOK, gin.H{
				"commits": []CommitResponse{},
				"count":   0,
			})
			return
		}
		c.JSON(http.StatusInternalServerError, ErrorResponse{
			Error: fmt.Sprintf("failed to get log: %v", err),
			Code:  "LOG_FAILED",
		})
		return
	}

	// Convert to response format
	commitResponses := make([]CommitResponse, len(commits))
	for i, commit := range commits {
		commitResponses[i] = CommitResponse{
			Hash:      commit.Hash(),
			Message:   commit.Message,
			Author:    commit.Author.Name,
			Email:     commit.Author.Email,
			Timestamp: commit.Author.Time,
			Parent:    commit.ParentHash,
		}
	}

	c.JSON(http.StatusOK, gin.H{
		"commits": commitResponses,
		"count":   len(commitResponses),
	})
}

// getStatus returns repository status
// @Summary Get repository status
// @Description Returns the current status of the working directory and staging area
// @Tags Git Operations
// @Produce json
// @Param repo_id path string true "Repository ID"
// @Success 200 {object} StatusResponse "Repository status"
// @Failure 404 {object} ErrorResponse "Repository not found"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /repositories/{repo_id}/status [get]
func (s *Server) getStatus(c *gin.Context) {
	repoID := c.Param("repo_id")
	repo, err := s.getRepository(repoID)
	if err != nil {
		c.JSON(http.StatusNotFound, ErrorResponse{
			Error: err.Error(),
			Code:  "REPO_NOT_FOUND",
		})
		return
	}

	status, err := repo.Status()
	if err != nil {
		c.JSON(http.StatusInternalServerError, ErrorResponse{
			Error: fmt.Sprintf("failed to get status: %v", err),
			Code:  "STATUS_FAILED",
		})
		return
	}

	// Calculate clean status
	isClean := len(status.Staged) == 0 && len(status.Modified) == 0 && len(status.Untracked) == 0

	// Create combined changes list for backward compatibility
	var changes []string
	changes = append(changes, status.Staged...)
	changes = append(changes, status.Modified...)
	changes = append(changes, status.Untracked...)

	// Return both the standard format and the "changes" format for backward compatibility
	c.JSON(http.StatusOK, gin.H{
		"branch":    status.Branch,
		"staged":    status.Staged,
		"modified":  status.Modified,
		"untracked": status.Untracked,
		"clean":     isClean,
		"changes":   changes,
	})
}

// showCommit returns details of a specific commit
// @Summary Get commit details
// @Description Retrieves detailed information about a specific commit including files and diff
// @Tags Git Operations
// @Produce json
// @Param repo_id path string true "Repository ID"
// @Param commit path string true "Commit hash"
// @Success 200 {object} object{commit=CommitResponse,tree=string,files=[]FileResponse,diff=object} "Commit details"
// @Failure 400 {object} ErrorResponse "Invalid request or not a commit"
// @Failure 404 {object} ErrorResponse "Repository or commit not found"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /repositories/{repo_id}/commits/{commit} [get]
func (s *Server) showCommit(c *gin.Context) {
	repoID := c.Param("repo_id")
	commitHash := c.Param("commit")

	repo, err := s.getRepository(repoID)
	if err != nil {
		c.JSON(http.StatusNotFound, ErrorResponse{
			Error: err.Error(),
			Code:  "REPO_NOT_FOUND",
		})
		return
	}

	// Get commit object
	obj, err := repo.GetObject(commitHash)
	if err != nil {
		c.JSON(http.StatusNotFound, ErrorResponse{
			Error: fmt.Sprintf("commit not found: %s", commitHash),
			Code:  "COMMIT_NOT_FOUND",
		})
		return
	}

	commit, ok := obj.(*object.Commit)
	if !ok {
		c.JSON(http.StatusBadRequest, ErrorResponse{
			Error: fmt.Sprintf("object %s is not a commit", commitHash),
			Code:  "NOT_A_COMMIT",
		})
		return
	}

	// Get tree for the commit
	tree, err := repo.GetObject(commit.TreeHash)
	if err != nil {
		c.JSON(http.StatusInternalServerError, ErrorResponse{
			Error: "failed to get commit tree",
			Code:  "TREE_NOT_FOUND",
		})
		return
	}

	treeObj, _ := tree.(*object.Tree)
	files := make([]FileResponse, 0, len(treeObj.Entries))
	for _, entry := range treeObj.Entries {
		files = append(files, FileResponse{
			Path: entry.Name,
			Size: 0, // Would need to fetch blob to get size
		})
	}

	// Get diff with parent commit
	var diffFiles []FileDiff
	if commit.ParentHash != "" {
		// Get parent and current trees to analyze changes
		parentObj, err := repo.GetObject(commit.ParentHash)
		if err == nil {
			parentCommit := parentObj.(*object.Commit)
			parentTree, _ := repo.GetObject(parentCommit.TreeHash)
			parentTreeObj := parentTree.(*object.Tree)

			// Compare trees to find changed files
			changedFiles := make(map[string]bool)

			// Check for removed/modified files from parent
			for _, entry := range parentTreeObj.Entries {
				changedFiles[entry.Name] = true
			}

			// Check for added/modified files in current
			for _, entry := range treeObj.Entries {
				changedFiles[entry.Name] = true
			}

			// Create file diff entries for changed files
			for path := range changedFiles {
				status := "modified"

				// Determine actual status
				inParent := false
				inCurrent := false

				for _, e := range parentTreeObj.Entries {
					if e.Name == path {
						inParent = true
						break
					}
				}

				for _, e := range treeObj.Entries {
					if e.Name == path {
						inCurrent = true
						break
					}
				}

				if !inParent && inCurrent {
					status = "added"
				} else if inParent && !inCurrent {
					status = "deleted"
				}

				diffFiles = append(diffFiles, FileDiff{
					Path:   path,
					Status: status,
				})
			}
		}
	}

	c.JSON(http.StatusOK, gin.H{
		"commit": CommitResponse{
			Hash:      commit.Hash(),
			Message:   commit.Message,
			Author:    commit.Author.Name,
			Email:     commit.Author.Email,
			Timestamp: commit.Author.Time,
			Parent:    commit.ParentHash,
		},
		"tree":  commit.TreeHash,
		"files": files,
		"diff": gin.H{
			"files": diffFiles,
		},
	})
}