# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A high-performance Git version control system API with memory-first architecture",
        "title": "GOVC API",
        "termsOfService": "https://github.com/caiatech/govc/blob/main/LICENSE",
        "contact": {
            "name": "GOVC API Support",
            "url": "https://github.com/caiatech/govc/issues",
            "email": "support@caiatech.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/caiatech/govc/blob/main/LICENSE"
        },
        "version": "1.0.0-dev"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/repositories/{repo_id}/commits": {
            "get": {
                "description": "Retrieves the commit history for a repository",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Git Operations"
                ],
                "summary": "Get commit history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository ID",
                        "name": "repo_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Maximum number of commits to return (default: 50, max: 10000)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commit history",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "commits": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/api.CommitResponse"
                                    }
                                },
                                "count": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new commit with staged changes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Git Operations"
                ],
                "summary": "Create a new commit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository ID",
                        "name": "repo_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Commit request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CommitRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Commit created successfully",
                        "schema": {
                            "$ref": "#/definitions/api.CommitResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/repositories/{repo_id}/commits/{commit}": {
            "get": {
                "description": "Retrieves detailed information about a specific commit including files and diff",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Git Operations"
                ],
                "summary": "Get commit details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository ID",
                        "name": "repo_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Commit hash",
                        "name": "commit",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commit details",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "commit": {
                                    "$ref": "#/definitions/api.CommitResponse"
                                },
                                "diff": {
                                    "type": "object"
                                },
                                "files": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/api.FileResponse"
                                    }
                                },
                                "tree": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request or not a commit",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Repository or commit not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/repositories/{repo_id}/files": {
            "post": {
                "description": "Writes a file to the working directory and stages it for commit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Git Operations"
                ],
                "summary": "Add a file to repository",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository ID",
                        "name": "repo_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "File addition request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.AddFileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File added successfully",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/repositories/{repo_id}/status": {
            "get": {
                "description": "Returns the current status of the working directory and staging area",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Git Operations"
                ],
                "summary": "Get repository status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository ID",
                        "name": "repo_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Repository status",
                        "schema": {
                            "$ref": "#/definitions/api.StatusResponse"
                        }
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v2/repositories/{repo_id}/commits": {
            "get": {
                "description": "Retrieves the commit history using V2 clean architecture with enhanced performance",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "V2 Architecture"
                ],
                "summary": "Get commit history (V2 Clean Architecture)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository ID",
                        "name": "repo_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Maximum number of commits to return (default: 20, max: 10000)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commit history",
                        "schema": {
                            "$ref": "#/definitions/api.LogResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new commit with staged changes using V2 clean architecture",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "V2 Architecture"
                ],
                "summary": "Create a new commit (V2 Clean Architecture)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository ID",
                        "name": "repo_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Commit request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CommitRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Commit created successfully",
                        "schema": {
                            "$ref": "#/definitions/api.CommitResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v2/repositories/{repo_id}/status": {
            "get": {
                "description": "Returns the current status using V2 clean architecture with improved accuracy",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "V2 Architecture"
                ],
                "summary": "Get repository status (V2 Clean Architecture)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository ID",
                        "name": "repo_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Repository status",
                        "schema": {
                            "$ref": "#/definitions/api.StatusResponse"
                        }
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.AddFileRequest": {
            "type": "object",
            "required": [
                "content",
                "path"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "example": "# My Project"
                },
                "path": {
                    "type": "string",
                    "example": "README.md"
                }
            }
        },
...
        "api.CommitRequest": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "author": {
                    "type": "string",
                    "example": "John Doe"
                },
                "email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "message": {
                    "type": "string",
                    "example": "Initial commit"
                }
            }
        },
        "api.CommitResponse": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string",
                    "example": "John Doe"
                },
                "email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "hash": {
                    "type": "string",
                    "example": "abc123def456"
                },
                "message": {
                    "type": "string",
                    "example": "Initial commit"
                },
                "parent": {
                    "type": "string",
                    "example": "def456abc123"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                }
            }
        },
...
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "REPO_NOT_FOUND"
                },
                "details": {},
                "error": {
                    "type": "string",
                    "example": "Repository not found"
                }
            }
        },
        "api.FileResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "api.LogResponse": {
            "type": "object",
            "properties": {
                "commits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.CommitResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "api.StatusResponse": {
            "type": "object",
            "properties": {
                "branch": {
                    "type": "string",
                    "example": "main"
                },
                "clean": {
                    "type": "boolean",
                    "example": false
                },
                "modified": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"file2.txt\"]"
                    ]
                },
                "staged": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"file1.txt\"]"
                    ]
                },
                "untracked": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"file3.txt\"]"
                    ]
                }
            }
        },
        "api.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Repository deleted successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Enter the token with the `Bearer: ` prefix, e.g. \"Bearer abcde12345\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ApiKeyQuery": {
            "description": "API key for accessing the API",
            "type": "apiKey",
            "name": "api_key",
            "in": "query"
        },
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}